Implement redux into Track component
