+++ src/components/presentation/App.js
@@ -1,9 +1,9 @@
 import {
   BrowserRouter, Switch, Route,
 } from 'react-router-dom';
-import LogIn from '../container/LogIn';
-import SignUp from '../container/SignUp';
-import Measurement from '../container/Measurement';
+import LogIn from './LogIn';
+import SignUp from './SignUp';
+import Measurement from './Measurement';
 import Track from '../container/Track';
 import TrackDetail from '../container/TrackDetail';
 import Progress from '../container/Progress';
